+++ DiploGM/models/player.py
@@ -1,10 +1,13 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Optional, Sequence
+from enum import Enum
+import discord
 
 from DiploGM.models import order
+from DiploGM.models.board import Board
+from DiploGM.utils import simple_player_name
 
-from enum import Enum
 
 if TYPE_CHECKING:
     from DiploGM.models import province
@@ -64,6 +67,17 @@ class Player:
         self.points: int = 0
         self.liege: Player | None = None
         self.vassals: list[Player] = []
+
+        # Must be initialised when the board is made
+        self.board = Optional[Board] = None
+
+
+    def find_discord_role(self, roles: Sequence[discord.Role]) -> Optional[discord.Role]:
+        for role in roles:
+            if simple_player_name(role.name) == simple_player_name(self.name):
+                return role
+        return None
+
 
     def __str__(self):
         return self.name